move unit setup
