--- data_functions.py
@@ -1,817 +1,0 @@
-current_user_info_document = {
-  "_id": "67456133140",
-  "language_of_instruction": "English",
-  "student_department": "Computer Engineering",
-  "student_faculty": "Faculty of Engineering / Computer Engineering",
-  "student_name": "EMİR ÇETİN",
-  "student_school_id": "041901027",
-  "student_status": "Continuing",
-  "student_surname": "MEMİŞ"
-}
-current_user_data_document = {
-  "owner_id": "67456133140",
-  "parsing_type": "offline",
-  "parsing_language": "en",
-  "transcript_manager_date": "20/05/2023 02:04:42",
-  "transcript_creation_date": "29.04.2023",
-  "semesters": {
-	"semester_1": {
-	  "semester_definition": "2020-2021 Fall Semester",
-	  "course_list": [
-		{
-		  "course_code": "COMP 100",
-		  "course_name": "Introduction to Computer Engineering",
-		  "course_lang": "EN",
-		  "course_credit": "3",
-		  "course_grade": "A",
-		  "course_grade_point": "12.00"
-		},
-		{
-		  "course_code": "COMP 109",
-		  "course_name": "Computer Programming (JAVA)",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "A",
-		  "course_grade_point": "24.00"
-		},
-		{
-		  "course_code": "ENG 101",
-		  "course_name": "English for Academic Purposes I",
-		  "course_lang": "EN",
-		  "course_credit": "4",
-		  "course_grade": "B",
-		  "course_grade_point": "12.00"
-		},
-		{
-		  "course_code": "MATH 115*",
-		  "course_name": "Calculus I",
-		  "course_lang": "EN",
-		  "course_credit": "7",
-		  "course_grade": "C+",
-		  "course_grade_point": "16.10"
-		},
-		{
-		  "course_code": "PHYS 103",
-		  "course_name": "Physics I",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "A",
-		  "course_grade_point": "24.00"
-		},
-		{
-		  "course_code": "PHYS 103L",
-		  "course_name": "Physics I Lab.",
-		  "course_lang": "EN",
-		  "course_credit": "2",
-		  "course_grade": "A",
-		  "course_grade_point": "8.00"
-		},
-		{
-		  "course_code": "TURK 111",
-		  "course_name": "Turkish Language and Literature I",
-		  "course_lang": "TR",
-		  "course_credit": "2",
-		  "course_grade": "A-",
-		  "course_grade_point": "7.40"
-		}
-	  ]
-	},
-	"semester_2": {
-	  "semester_definition": "2020-2021 Spring Semester",
-	  "course_list": [
-		{
-		  "course_code": "COMP 110",
-		  "course_name": "Object-Oriented Programming (JAVA)",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "A",
-		  "course_grade_point": "24.00"
-		},
-		{
-		  "course_code": "ENG 102",
-		  "course_name": "English for Academic Purposes II",
-		  "course_lang": "EN",
-		  "course_credit": "4",
-		  "course_grade": "A-",
-		  "course_grade_point": "14.80"
-		},
-		{
-		  "course_code": "MATH 108",
-		  "course_name": "Discrete and Combinatorial Mathematics",
-		  "course_lang": "EN",
-		  "course_credit": "5",
-		  "course_grade": "C+",
-		  "course_grade_point": "11.50"
-		},
-		{
-		  "course_code": "MATH 116",
-		  "course_name": "Calculus II",
-		  "course_lang": "EN",
-		  "course_credit": "7",
-		  "course_grade": "A",
-		  "course_grade_point": "28.00"
-		},
-		{
-		  "course_code": "PHYS 104",
-		  "course_name": "Physics II",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "A",
-		  "course_grade_point": "24.00"
-		},
-		{
-		  "course_code": "PHYS 104L",
-		  "course_name": "Physics II Lab",
-		  "course_lang": "EN",
-		  "course_credit": "2",
-		  "course_grade": "A",
-		  "course_grade_point": "8.00"
-		}
-	  ]
-	},
-	"semester_3": {
-	  "semester_definition": "2021-2022 Fall Semester",
-	  "course_list": [
-		{
-		  "course_code": "COMP 201",
-		  "course_name": "Data Structures and Algorithms",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "A",
-		  "course_grade_point": "24.00"
-		},
-		{
-		  "course_code": "COMP 205",
-		  "course_name": "Systems Programming",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "A",
-		  "course_grade_point": "24.00"
-		},
-		{
-		  "course_code": "EE 203*",
-		  "course_name": "Digital Systems Design",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "B-",
-		  "course_grade_point": "16.20"
-		},
-		{
-		  "course_code": "ENGR 301",
-		  "course_name": "Technical Report Writing and Presentation",
-		  "course_lang": "EN",
-		  "course_credit": "4",
-		  "course_grade": "A-",
-		  "course_grade_point": "14.80"
-		},
-		{
-		  "course_code": "HISTR 211*",
-		  "course_name": "Principles of Ataturk and History of the Turkish Republic I",
-		  "course_lang": "TR",
-		  "course_credit": "2",
-		  "course_grade": "B+",
-		  "course_grade_point": "6.60"
-		},
-		{
-		  "course_code": "MATH 115*",
-		  "course_name": "Calculus I",
-		  "course_lang": "EN",
-		  "course_credit": "7",
-		  "course_grade": "C+",
-		  "course_grade_point": "16.10"
-		},
-		{
-		  "course_code": "MATH 211",
-		  "course_name": "Linear Algebra",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "A-",
-		  "course_grade_point": "22.20"
-		}
-	  ]
-	},
-	"semester_4": {
-	  "semester_definition": "2021-2022 Spring Semester",
-	  "course_list": [
-		{
-		  "course_code": "COMP 204",
-		  "course_name": "Programming Studio",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "A",
-		  "course_grade_point": "24.00"
-		},
-		{
-		  "course_code": "COMP 206",
-		  "course_name": "Computer Architecture",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "C",
-		  "course_grade_point": "12.00"
-		},
-		{
-		  "course_code": "EE 212",
-		  "course_name": "Electrical and Electronic Circuits",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "C-",
-		  "course_grade_point": "10.20"
-		},
-		{
-		  "course_code": "GE 201",
-		  "course_name": "Personalized Learning Experience I",
-		  "course_lang": "EN",
-		  "course_credit": "5",
-		  "course_grade": "S",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "HISTR 211",
-		  "course_name": "Principles of Ataturk and History of the Turkish Republic I",
-		  "course_lang": "TR",
-		  "course_credit": "2",
-		  "course_grade": "A-",
-		  "course_grade_point": "7.40"
-		},
-		{
-		  "course_code": "HISTR 212",
-		  "course_name": "Principles of Ataturk and History of the Turkish Republic II",
-		  "course_lang": "TR",
-		  "course_credit": "2",
-		  "course_grade": "A-",
-		  "course_grade_point": "7.40"
-		},
-		{
-		  "course_code": "MATH 115*",
-		  "course_name": "Calculus I",
-		  "course_lang": "EN",
-		  "course_credit": "7",
-		  "course_grade": "F",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "MATH 224",
-		  "course_name": "Probability and Statistics for Engineering",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "C-",
-		  "course_grade_point": "10.20"
-		}
-	  ]
-	},
-	"semester_5": {
-	  "semester_definition": "2022-2023 Fall Semester",
-	  "course_list": [
-		{
-		  "course_code": "COMP 303",
-		  "course_name": "Analysis of Algorithms",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "B-",
-		  "course_grade_point": "16.20"
-		},
-		{
-		  "course_code": "COMP 305",
-		  "course_name": "Database Management Systems",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "C",
-		  "course_grade_point": "12.00"
-		},
-		{
-		  "course_code": "EE 203",
-		  "course_name": "Digital Systems Design",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "A",
-		  "course_grade_point": "24.00"
-		},
-		{
-		  "course_code": "GE 202",
-		  "course_name": "Personalized Learning Experience II",
-		  "course_lang": "EN",
-		  "course_credit": "5",
-		  "course_grade": "S",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "MATH 115",
-		  "course_name": "Calculus I",
-		  "course_lang": "EN",
-		  "course_credit": "7",
-		  "course_grade": "F",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "MATH 213*",
-		  "course_name": "Differential Equations",
-		  "course_lang": "EN",
-		  "course_credit": "7",
-		  "course_grade": "W",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "MATH 321",
-		  "course_name": "Automata Theory and Formal Language",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "W",
-		  "course_grade_point": "0.00"
-		}
-	  ]
-	},
-	"semester_6": {
-	  "semester_definition": "2022-2023 Spring Semester",
-	  "course_list": [
-		{
-		  "course_code": "COMP 302",
-		  "course_name": "Software Engineering",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "N/A",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "COMP 304",
-		  "course_name": "Operating Systems",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "N/A",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "ECON 112",
-		  "course_name": "Economics for Engineering",
-		  "course_lang": "EN",
-		  "course_credit": "5",
-		  "course_grade": "N/A",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "EE 204",
-		  "course_name": "Signals and Systems",
-		  "course_lang": "EN",
-		  "course_credit": "6",
-		  "course_grade": "N/A",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "EE 306",
-		  "course_name": "Microprocessors",
-		  "course_lang": "EN",
-		  "course_credit": "7",
-		  "course_grade": "N/A",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "MATH 213",
-		  "course_name": "Differential Equations",
-		  "course_lang": "EN",
-		  "course_credit": "7",
-		  "course_grade": "N/A",
-		  "course_grade_point": "0.00"
-		},
-		{
-		  "course_code": "TURK 112",
-		  "course_name": "Turkish Language and Literature II",
-		  "course_lang": "TR",
-		  "course_credit": "2",
-		  "course_grade": "N/A",
-		  "course_grade_point": "0.00"
-		}
-	  ]
-	}
-  },
-  "original_course_list": [
-	{
-	  "course_code": "COMP 100",
-	  "course_name": "Introduction to Computer Engineering",
-	  "course_lang": "EN",
-	  "course_credit": "3",
-	  "course_grade": "A",
-	  "course_grade_point": "12.00"
-	},
-	{
-	  "course_code": "COMP 109",
-	  "course_name": "Computer Programming (JAVA)",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "A",
-	  "course_grade_point": "24.00"
-	},
-	{
-	  "course_code": "ENG 101",
-	  "course_name": "English for Academic Purposes I",
-	  "course_lang": "EN",
-	  "course_credit": "4",
-	  "course_grade": "B",
-	  "course_grade_point": "12.00"
-	},
-	{
-	  "course_code": "PHYS 103",
-	  "course_name": "Physics I",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "A",
-	  "course_grade_point": "24.00"
-	},
-	{
-	  "course_code": "PHYS 103L",
-	  "course_name": "Physics I Lab.",
-	  "course_lang": "EN",
-	  "course_credit": "2",
-	  "course_grade": "A",
-	  "course_grade_point": "8.00"
-	},
-	{
-	  "course_code": "TURK 111",
-	  "course_name": "Turkish Language and Literature I",
-	  "course_lang": "TR",
-	  "course_credit": "2",
-	  "course_grade": "A-",
-	  "course_grade_point": "7.40"
-	},
-	{
-	  "course_code": "COMP 110",
-	  "course_name": "Object-Oriented Programming (JAVA)",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "A",
-	  "course_grade_point": "24.00"
-	},
-	{
-	  "course_code": "ENG 102",
-	  "course_name": "English for Academic Purposes II",
-	  "course_lang": "EN",
-	  "course_credit": "4",
-	  "course_grade": "A-",
-	  "course_grade_point": "14.80"
-	},
-	{
-	  "course_code": "MATH 108",
-	  "course_name": "Discrete and Combinatorial Mathematics",
-	  "course_lang": "EN",
-	  "course_credit": "5",
-	  "course_grade": "C+",
-	  "course_grade_point": "11.50"
-	},
-	{
-	  "course_code": "MATH 116",
-	  "course_name": "Calculus II",
-	  "course_lang": "EN",
-	  "course_credit": "7",
-	  "course_grade": "A",
-	  "course_grade_point": "28.00"
-	},
-	{
-	  "course_code": "PHYS 104",
-	  "course_name": "Physics II",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "A",
-	  "course_grade_point": "24.00"
-	},
-	{
-	  "course_code": "PHYS 104L",
-	  "course_name": "Physics II Lab",
-	  "course_lang": "EN",
-	  "course_credit": "2",
-	  "course_grade": "A",
-	  "course_grade_point": "8.00"
-	},
-	{
-	  "course_code": "COMP 201",
-	  "course_name": "Data Structures and Algorithms",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "A",
-	  "course_grade_point": "24.00"
-	},
-	{
-	  "course_code": "COMP 205",
-	  "course_name": "Systems Programming",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "A",
-	  "course_grade_point": "24.00"
-	},
-	{
-	  "course_code": "ENGR 301",
-	  "course_name": "Technical Report Writing and Presentation",
-	  "course_lang": "EN",
-	  "course_credit": "4",
-	  "course_grade": "A-",
-	  "course_grade_point": "14.80"
-	},
-	{
-	  "course_code": "MATH 211",
-	  "course_name": "Linear Algebra",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "A-",
-	  "course_grade_point": "22.20"
-	},
-	{
-	  "course_code": "COMP 204",
-	  "course_name": "Programming Studio",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "A",
-	  "course_grade_point": "24.00"
-	},
-	{
-	  "course_code": "COMP 206",
-	  "course_name": "Computer Architecture",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "C",
-	  "course_grade_point": "12.00"
-	},
-	{
-	  "course_code": "EE 212",
-	  "course_name": "Electrical and Electronic Circuits",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "C-",
-	  "course_grade_point": "10.20"
-	},
-	{
-	  "course_code": "GE 201",
-	  "course_name": "Personalized Learning Experience I",
-	  "course_lang": "EN",
-	  "course_credit": "5",
-	  "course_grade": "S",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "HISTR 211",
-	  "course_name": "Principles of Ataturk and History of the Turkish Republic I",
-	  "course_lang": "TR",
-	  "course_credit": "2",
-	  "course_grade": "A-",
-	  "course_grade_point": "7.40"
-	},
-	{
-	  "course_code": "HISTR 212",
-	  "course_name": "Principles of Ataturk and History of the Turkish Republic II",
-	  "course_lang": "TR",
-	  "course_credit": "2",
-	  "course_grade": "A-",
-	  "course_grade_point": "7.40"
-	},
-	{
-	  "course_code": "MATH 224",
-	  "course_name": "Probability and Statistics for Engineering",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "C-",
-	  "course_grade_point": "10.20"
-	},
-	{
-	  "course_code": "COMP 303",
-	  "course_name": "Analysis of Algorithms",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "B-",
-	  "course_grade_point": "16.20"
-	},
-	{
-	  "course_code": "COMP 305",
-	  "course_name": "Database Management Systems",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "C",
-	  "course_grade_point": "12.00"
-	},
-	{
-	  "course_code": "EE 203",
-	  "course_name": "Digital Systems Design",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "A",
-	  "course_grade_point": "24.00"
-	},
-	{
-	  "course_code": "GE 202",
-	  "course_name": "Personalized Learning Experience II",
-	  "course_lang": "EN",
-	  "course_credit": "5",
-	  "course_grade": "S",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "MATH 115",
-	  "course_name": "Calculus I",
-	  "course_lang": "EN",
-	  "course_credit": "7",
-	  "course_grade": "F",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "MATH 321",
-	  "course_name": "Automata Theory and Formal Language",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "W",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "COMP 302",
-	  "course_name": "Software Engineering",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "N/A",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "COMP 304",
-	  "course_name": "Operating Systems",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "N/A",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "ECON 112",
-	  "course_name": "Economics for Engineering",
-	  "course_lang": "EN",
-	  "course_credit": "5",
-	  "course_grade": "N/A",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "EE 204",
-	  "course_name": "Signals and Systems",
-	  "course_lang": "EN",
-	  "course_credit": "6",
-	  "course_grade": "N/A",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "EE 306",
-	  "course_name": "Microprocessors",
-	  "course_lang": "EN",
-	  "course_credit": "7",
-	  "course_grade": "N/A",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "MATH 213",
-	  "course_name": "Differential Equations",
-	  "course_lang": "EN",
-	  "course_credit": "7",
-	  "course_grade": "N/A",
-	  "course_grade_point": "0.00"
-	},
-	{
-	  "course_code": "TURK 112",
-	  "course_name": "Turkish Language and Literature II",
-	  "course_lang": "TR",
-	  "course_credit": "2",
-	  "course_grade": "N/A",
-	  "course_grade_point": "0.00"
-	}
-  ],
-  "filtering": None,
-  "sorting": None,
-  "modified_course_list": None,
-  "document_name": None,
-  "subtracted_course_list": None,
-  "added_course_list": None
-}
-
-language_of_instruction = current_user_info_document["language_of_instruction"]
-student_department = current_user_info_document["student_department"]
-student_faculty = current_user_info_document["student_faculty"]
-student_name = current_user_info_document["student_name"]
-student_school_id = current_user_info_document["student_school_id"]
-student_national_id = current_user_info_document["_id"]
-student_status = current_user_info_document["student_status"]
-student_surname = current_user_info_document["student_surname"]
-
-owner_id = current_user_data_document["owner_id"]
-parsing_type = current_user_data_document["parsing_type"]
-parsing_language = current_user_data_document["parsing_language"]
-transcript_manager_date = current_user_data_document["transcript_manager_date"]
-transcript_creation_date = current_user_data_document["transcript_creation_date"]
-semesters = current_user_data_document["semesters"]
-original_course_list = current_user_data_document["original_course_list"]
-filtering = current_user_data_document["filtering"]
-sorting = current_user_data_document["sorting"]
-modified_course_list = current_user_data_document["modified_course_list"]
-document_name = current_user_data_document["document_name"]
-subtracted_course_list = current_user_data_document["subtracted_course_list"]
-added_course_list = current_user_data_document["added_course_list"]
-
-def sort_by(given_course_list, sort_key, should_reverse=False):
-	course_list = given_course_list.copy()
-	course_list.sort(key=lambda x: x[sort_key], reverse=should_reverse)
-	return course_list
-
-def filter_by(given_course_list, filter_key, filter_value):
-	course_list = given_course_list.copy()
-	course_list = list(filter(lambda x: x[filter_key] == filter_value, course_list))
-	return course_list
-
-def add_course(given_course_list, course_code, course_name, course_lang, course_credit, course_grade, course_grade_point):
-	course_list = given_course_list.copy()
-	result = filter_by(course_list, "course_code", course_code)
-	n = 1
-	while result:
-		course_code = course_code + " (" + str(n) + ")"
-		result = filter_by(course_list, "course_code", course_code)
-		n += 1
-	new_course ={
-		"course_code": course_code,
-		"course_name": course_name,
-		"course_lang": course_lang,
-		"course_credit": course_credit,
-		"course_grade": course_grade,
-		"course_grade_point": course_grade_point
-	}
-	course_list.append(new_course)
-	return course_list
-
-def subtract_course(given_course_list, course_code):
-	course_list = given_course_list.copy()
-	result = filter_by(course_list, "course_code", course_code)
-	if not result:
-		return course_list
-	else:
-		course_list = list(filter(lambda x: x["course_code"] != course_code, course_list))
-		return course_list
-
-def update_course(given_course_list, course_code, course_name, course_lang, course_credit, course_grade, course_grade_point):
-	# The code is defining a function that takes a list of dictionaries as input and filters the list
-	# based on a given course code. If the course code is not found in the list, the original list is
-	# returned. If the course code is found, the function removes the dictionary with that course code
-	# from the list and adds a new dictionary with the given course information. The updated list is
-	# then returned.
-	course_list = given_course_list.copy()
-	result = filter_by(course_list, "course_code", course_code)
-	if not result:
-		return course_list
-	else:
-		course_list = list(filter(lambda x: x["course_code"] != course_code, course_list))
-		course_list = add_course(course_list, course_code, course_name, course_lang, course_credit, course_grade, course_grade_point)
-		return course_list
-
-def calculate_gpa(course_list):
-	"""
-	This function calculates the GPA based on the given course list.
-	"""
-
-	credits_attempted = 0
-	credits_successful = 0
-	credits_included_in_gpa = 0
-	gpa = 0
-
-	for course in course_list:
-		course_credit : int = int(course["course_credit"])
-		course_grade : str = course["course_grade"]
-		course_grade_point : float = float(course["course_grade_point"])
-
-		credits_attempted += course_credit
-
-		if course_grade in ["A", "A-", "B+", "B", "B-", "C+", "C", "C-", "D+", "D"]:
-			credits_successful += course_credit
-			credits_included_in_gpa += course_credit
-			gpa += course_grade_point
-		elif course_grade in ["S"]:
-			credits_successful += course_credit
-		elif course_grade == "F" or course_grade == "U":
-			credits_included_in_gpa += course_credit
-			gpa += course_grade_point
-		elif course_grade == "I":
-			continue
-		elif course_grade == "W":
-			credits_attempted -= course_credit
-		elif course_grade == "N/A":
-			continue
-
-	if credits_included_in_gpa == 0:
-		gpa = 0
-	else :
-		gpa = gpa / credits_included_in_gpa
-	
-	gpa = round(gpa, 2)
-
-	return (
-		credits_attempted,
-		credits_successful,
-		credits_included_in_gpa,
-		gpa
-	)
-
-
-course_list = semesters["semester_1"]["course_list"]
-
-course_list = sort_by(course_list, "course_code")
-course_list = sort_by(course_list, "course_grade")
-course_list = sort_by(course_list, "course_grade", should_reverse=True)
-
-course_list = semesters["semester_5"]["course_list"]
-
-course_list = filter_by(course_list, "course_lang", "EN")
-course_list_2 = filter_by(course_list, "course_grade", "A")
-
